backend/pyproject.toml
[tool.poetry]
name = "connect-four-backend"
version = "0.1.0"
description = ""
authors = ["feaad"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"
uwsgi = "^2.0.26"
psycopg = {extras = ["binary"], version = "^3.2.1"}
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
celery = "^5.4.0"
redis = "^5.0.7"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
numpy = "^2.0.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
